updated files

+++ src/practice_package/roller_coaster.py
@@ -1,7 +1,7 @@
 height = float(input("Enter your height: "))
 if height > 4.0:
     print("You are eligible to ride")
-    age = float(input("How old are you: "))
+    age = int(input("How old are you: "))
     if 11 < age < 61:
         print("Kindly pay #200 to ride")
     elif age < 12:
